Update License
2012 -> 2013

Fix the bug of resized box event.

After the provider get the resized event from the box, need to check
the current size of a box.

If is same with requested size, send it to the client that the request
is already finished.

Apply the livebox errno (status code) instead of system errno.
Make easy to understand what happens in the API.

Use the util_uri_to_path to unlink image files of liveboxes(image type)

Change-Id: Ide9281da21700f4e1d9c0eb480daeff891258ccd