Care the case for failing open a PD.
If a provider fails to load script for PD,
It should notice to slave provider.

Currently, the slave will prevent other boxes update if the PD
is opened of a box.
So, if the master doesn't send the failed event to the slave,
and only client notice it, the slave cannot get any failed or close PD event.
then it will block all other boxes update.

This patch will fix it.

Now if the master fails to load the EDJE for PD,
the slave will notice it via close,pd event.
then the other box can be updated normally.

Change-Id: I44f7bf6378e07adc29656db94993f9af38a71d34