Fix a bug that the instance is not deleted correctly
in specific cases, the master will not release the deleted instance'd resources.
It could be a leak.

After open a PD, the timer will be registered and the instance's refcnt will be increased.
But even though the service provider makes response via acquire_buffer, the master
doesn't decrease the instance's refcnt.
It makes a problem. even if the instance is deleted, the livebox doesn't cleared.
because the refcnt could not be reach to ZERO.

This patch will fix it.

And optimize the log.
Verbose logging message are deleted and optimized.

Change-Id: I1c85f428428ac16459c91dfe10dbfa43c972ed5b