Manage the fault of service providers(osp,web).
While processing the PD open/close operation, the service provider can be faulted.
In that case the master should care the processed operation.
So the master will add the slave deactivate event handler and if it is raised up,
the master will clean all resources related with monitoring PD open/close operations.

Change-Id: Ia5e9937d66af3192c4ff08ed26e923438627bfda